refactor simple game server
